correct the naming of routes to match requirements of theteacher  instructions
