--- src/Controller/Front/BookingController.php
@@ -1,83 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller\Front;
-
-use App\Entity\Property;
-use App\Entity\User;
-use App\Exception\UnavailableDatesException;
-use App\Exception\BookingConflictException;
-use App\Form\BookingType;
-use App\Repository\PropertyRepository;
-use App\Service\BookingService;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
-use App\Security\Voter\PropertyVoter;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
-
-#[IsGranted('ROLE_USER')]
-final class BookingController extends AbstractController
-{
-    #[Route('/logement/{id}/reserver', name: 'app_booking_checkout', methods: ['GET', 'POST'])]
-    #[IsGranted(PropertyVoter::VIEW, subject: 'property')]
-    public function checkout(
-        Request $request,
-        Property $property,
-        PropertyRepository $propertyRepository,
-        BookingService $bookingService,
-    ): Response {
-        if ($property->getStatus() !== 'published') {
-            throw $this->createNotFoundException('Ce logement n\'est pas disponible à la réservation.');
-        }
-
-        $property = $propertyRepository->findOneForDetail($property) ?? $property;
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            return $this->redirectToRoute('app_login');
-        }
-
-        if ($property->getHost()?->getId() === $user->getId()) {
-            $this->addFlash('error', 'Vous ne pouvez pas réserver votre propre logement.');
-
-            return $this->redirectToRoute('app_logement_detail', ['id' => $property->getId()]);
-        }
-
-        $form = $this->createForm(BookingType::class);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $data = $form->getData();
-            $checkin = $data['checkinDate'];
-            $checkout = $data['checkoutDate'];
-            $guestsCount = (int) $data['guestsCount'];
-
-            try {
-                $reservation = $bookingService->create($property, $user, $checkin, $checkout, $guestsCount);
-            } catch (BookingConflictException|UnavailableDatesException|\LogicException $exception) {
-                $this->addFlash('error', $exception->getMessage());
-
-                return $this->render('front/property/booking.html.twig', [
-                    'property' => $property,
-                    'form' => $form,
-                ], new Response(status: Response::HTTP_UNPROCESSABLE_ENTITY));
-            }
-
-            $message = $reservation->getStatus() === 'confirmed'
-                ? 'Votre réservation est confirmée.'
-                : 'Votre demande de réservation a été transmise à l’hôte.';
-            $this->addFlash('success', $message);
-
-            return $this->redirectToRoute('app_reservation_show', ['id' => $reservation->getId()]);
-        }
-
-        return $this->render('front/property/booking.html.twig', [
-            'property' => $property,
-            'form' => $form,
-        ], new Response(
-            status: $form->isSubmitted() ? Response::HTTP_UNPROCESSABLE_ENTITY : Response::HTTP_OK
-        ));
-    }
-}
