+++ src/Controller/Front/ReservationController.php
@@ -7,12 +7,12 @@ namespace App\Controller\Front;
 use App\Entity\Reservation;
 use App\Entity\User;
 use App\Repository\ReservationRepository;
+use App\Security\Voter\ReservationVoter;
 use App\Service\BookingService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use App\Security\Voter\ReservationVoter;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/reservations')]
@@ -34,18 +34,9 @@ final class ReservationController extends AbstractController
 
     #[Route('/{id}', name: 'app_reservation_show', methods: ['GET'])]
     #[IsGranted(ReservationVoter::VIEW, subject: 'reservation')]
-    public function show(Reservation $reservation, ReservationRepository $reservationRepository): Response
+    public function show(Reservation $reservation): Response
     {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            return $this->redirectToRoute('app_login');
-        }
-
-        $reservation = $reservationRepository->findOneForDetail($reservation) ?? $reservation;
-
-        return $this->render('front/reservation/show.html.twig', [
-            'reservation' => $reservation,
-        ]);
+        return $this->redirectToRoute('app_booking_show', ['id' => $reservation->getId()]);
     }
 
     #[Route('/{id}/cancel', name: 'app_reservation_cancel', methods: ['POST'])]
@@ -55,7 +46,7 @@ final class ReservationController extends AbstractController
         if (!$this->isCsrfTokenValid('cancel_guest'.$reservation->getId(), $request->request->get('_token'))) {
             $this->addFlash('error', 'Jeton CSRF invalide.');
 
-            return $this->redirectToRoute('app_reservation_show', ['id' => $reservation->getId()]);
+            return $this->redirectToRoute('app_booking_show', ['id' => $reservation->getId()]);
         }
 
         $user = $this->getUser();
@@ -72,6 +63,6 @@ final class ReservationController extends AbstractController
             $this->addFlash('error', $exception->getMessage());
         }
 
-        return $this->redirectToRoute('app_reservation_show', ['id' => $reservation->getId()]);
+        return $this->redirectToRoute('app_booking_show', ['id' => $reservation->getId()]);
     }
 }
