+++ templates/front/host_calendar/index.html.twig
@@ -150,7 +150,7 @@
                                     {% if block.notes %}
                                         <p class="text-xs text-gray-500 mt-1">{{ block.notes }}</p>
                                     {% endif %}
-                                    <form action="{{ path('app_host_block_delete', { id: block.id }) }}" method="POST" class="mt-3">
+                                    <form action="{{ path('app_host_block_delete', { blockId: block.id }) }}" method="POST" class="mt-3">
                                         <input type="hidden" name="_token" value="{{ csrf_token('delete_block' ~ block.id) }}">
                                         <button type="submit" class="inline-flex items-center gap-2 text-xs font-semibold text-red-600 hover:text-red-700">
                                             <i class="fa-solid fa-trash"></i> Supprimer
